🎥🌇 Major refactor to MVC. More to come, then tests.
